Replaced Camera concept with Photo and Sources
It has been a while since I worked on this so I'm not certain, but it
seems that I ripped out almost everything to do with Camera and am just
using the exif data now and a small lookup table for nice names. In
addition, Photo is a more universal concept.